Moved tests to separate folder
